Modified all malloc/free/realloc calls in croaring to be gc-aware
This is an attempt to get the guile GC to manage the native objects
as opposed to having to manage the bitmap memory outside of guile via
finalizers. The one unknown here is how to handle calls to
posix_memalign, as there is no scm_gc_* equivalent. For the meantime
the size_t of the memalign'd memory is registered; but this is a leak
at the moment and will be fixed.